requirements.txt
flask==2.2.3
python-dotenv==0.19.0
requests==2.26.0
anthropic>=0.18.0
httpx>=0.24.1
python-dateutil==2.8.2
gunicorn==20.1.0
werkzeug==2.2.3
APScheduler==3.10.4
Flask-SQLAlchemy==3.0.3
sendgrid==6.10.0
twilio==8.12.0
pandas==2.0.3
openpyxl==3.1.2
PyPDF2==3.0.1
pdfplumber==0.9.0
python-pptx==0.6.21
chardet==5.2.0
Pillow==10.3.0
